Replace syntax code with Helix's
