+++ defaults/light-theme.toml
@@ -52,11 +52,17 @@ magenta = "#A626A4"
 "variable.other.member" = "$red"
 
 "string" = "$green"
+"string.escape" = "$cyan"
 
 "type.builtin" = "$cyan"
 "builtinType" = "$cyan"
 "escape" = "$cyan"
 "embedded" = "$cyan"
+
+"punctuation.delimiter" = "$yellow"
+"text.title" = "$orange"
+"text.uri" = "$cyan"
+"text.reference" = "$yellow"
 
 [theme.ui]
 "lapce.active_tab" = "$white"
