+++ lapce-core/Cargo.toml
@@ -9,6 +9,9 @@ thiserror = "1.0"
 itertools = "0.10.3"
 log = "0.4.14"
 bitflags = "1.3.2"
+once_cell = "1.13.0"
+slotmap = "1.0"
+arc-swap = "1.5.1"
 strum = "0.24.0"
 strum_macros = "0.24"
 serde = "1.0"
@@ -30,7 +33,7 @@ tree-sitter-ruby = { git = "https://github.com/Liberatys/tree-sitter-ruby.git", 
 tree-sitter-c = { version = "0.20.1", optional = true }
 tree-sitter-cpp = { version = "0.20.0", optional = true }
 tree-sitter-json = { version = "0.19.0", optional = true }
-tree-sitter-md = { git = "https://github.com/dzhou121/tree-sitter-markdown.git", version = "0.0.1", optional = true  }
+tree-sitter-md = { git = "https://github.com/MDeiml/tree-sitter-markdown.git", branch = "split_parser", optional = true }
 tree-sitter-html = { version = "0.19.0", optional = true }
 tree-sitter-java = { git = "https://github.com/tree-sitter/tree-sitter-java.git", version = "0.20.0", optional = true  }
 tree-sitter-elm = { version = "5.6.0", optional = true }
