+++ lapce-core/queries/markdown.inline/injections.scm
@@ -1,0 +1,2 @@
+
+((html_tag) @injection.content (#set! injection.language "html"))
