+++ lapce-core/src/word.rs
@@ -1,6 +1,6 @@
 use xi_rope::{Cursor, Rope, RopeInfo};
 
-use crate::syntax::{matching_char, matching_pair_direction};
+use crate::syntax::util::{matching_char, matching_pair_direction};
 
 #[derive(Copy, Clone, PartialEq, Eq)]
 pub enum WordProperty {
